feat: change fee mechanism, initial dummy commit
